Added an actual questionnaire page
Instead of doing individual pages, just use rounds and change the decision tree. The list of decision trees is listed in constants (C.TREE_ANSWERS). Each round a new decision tree is chosen. The correct answer associated with the tree is in the list as well so after every choice, the app checks whether the answer was correct or not.

Need to
1) Add attention checks
2) Add actual trees
3) Add the validation function and the total function at the end